OutOfService

+++ src/main/java/com/CSMS/CSMS/models/Booking.java
@@ -22,7 +22,7 @@ public class Booking {
 
     //Include User Id
 
-    private int customer_id;
+    private Long customer_id;
     private String start_time;
     private String end_time;
     private int cost;
@@ -36,7 +36,7 @@ public class Booking {
     private String bookingStatus;
     // private String chargerName;
 
-    public Booking(String bookingStatus,int customer_id, String booking_id, int station_id, int charger_id, int connector_id, String start_time, String end_time, int cost, String payment_mode, int payment_status, String date) {
+    public Booking(String bookingStatus,Long customer_id, String booking_id, int station_id, int charger_id, int connector_id, String start_time, String end_time, int cost, String payment_mode, int payment_status, String date) {
         this.customer_id = customer_id;
         this.booking_id = booking_id;
         this.station_id = station_id;
@@ -60,11 +60,11 @@ public class Booking {
         this.id = id;
     }
 
-    public int getCustomer_id() {
+    public Long getCustomer_id() {
         return customer_id;
     }
 
-    public void setCustomer_id(int customer_id) {
+    public void setCustomer_id(Long customer_id) {
         this.customer_id = customer_id;
     }
